+++ force-app/main/default/lwc/appArea/appArea.js-meta.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata" fqn="appArea">
+    <apiVersion>45.0</apiVersion>
+    <isExposed>true</isExposed>
+    <targets>
+	<target>lightning__AppPage</target>
+	<target>lightning__RecordPage</target>
+	<target>lightning__HomePage</target>
+</targets>
+</LightningComponentBundle>
