+++ force-app/main/default/lwc/appProductSearch/appProductSearch.html
@@ -1,15 +1,5 @@
 <template>
     <lightning-card>
-    <div >
-        <lightning-combobox
-            name='category'
-            label='Area'
-            placeholder='Select an area to add app too'
-            value={area}
-            options={areaOptions}
-            onchange={selectArea}></lightning-combobox>
-    </div>  
-    <br/> 
     <h4>Add Product</h4>
     <div class='slds-grid slds-gutters'>
         <div class="slds-col slds-size_5-of-12">
